hal_embedded_mcp/hal_embedded_mcp/server/requirements.txt
# HAL Embedded MCP Server v1 – talk to MCU over serial, expose tools to MCP clients
mcp>=1.0.0
pyserial>=3.5
